+++ tests/testthat/test-dataset_serialize.R
@@ -5,7 +5,7 @@ test_that("serializing dataset works", {
     x <- c("S", "P", "Q", "R")
     file <- tempfile()
     writeLines(paste0('"', x, '"'), file)
-    ds <- read_json(file)
+    ds <- read_ndjson(file)
 
     file2 <- tempfile()
     saveRDS(ds, file2)
@@ -35,7 +35,7 @@ test_that("serializing dataset should use relative, not absolute path", {
     # save dir/a/data.json
     # save dir/a/obj.rds
     writeLines(paste0('{"x": "', x, '"}'), "data.json")
-    ds <- read_json("data.json")
+    ds <- read_ndjson("data.json")
     saveRDS(ds, "obj.rds")
     rm("ds"); gc()
 
@@ -63,7 +63,7 @@ test_that("serializing dataset subset works", {
     x <- LETTERS
     file <- tempfile()
     writeLines(paste0('"', x, '"'), file)
-    ds <- read_json(file)
+    ds <- read_ndjson(file)
 
     i <- seq(2, 26, 2)
     ds <- ds[i]
@@ -85,7 +85,7 @@ test_that("serializing dataset field works", {
     y <- 3.14 * seq_along(LETTERS) - 10
     file <- tempfile()
     writeLines(paste0('{"x": "', x, '", "z": { "y": ', y, "} }"), file)
-    ds <- read_json(file)
+    ds <- read_ndjson(file)
 
     ds <- ds$z
 
@@ -105,7 +105,7 @@ test_that("serializing dataset nested fields works", {
     x <- 1:10
     file <- tempfile()
     writeLines(paste0('{"f1": {"f2": {"f3": {"x": ', x, '}}}}'), file)
-    ds <- read_json(file)
+    ds <- read_ndjson(file)
 
     ds <- ds$f1$f2$f3
 
@@ -126,7 +126,7 @@ test_that("serializing dataset field subset works", {
     y <- 3.14 * seq_along(LETTERS) - 10
     file <- tempfile()
     writeLines(paste0('{"z": {"x": "', x, '"}, "y": ', y, "}"), file)
-    ds <- read_json(file)
+    ds <- read_ndjson(file)
 
     ds <- ds$z
     i <- c(20, 2, 9, 4, 6, 2)